modificaciones generales

+++ app/Http/Controllers/ProductoController.php
@@ -21,49 +21,59 @@ class ProductoController extends Controller
     }
 
     //agrega un nuevo producto
-    public function guardarProducto(Request $request)
-{
-    // Validar la solicitud
-    $request->validate([
-        'marca_id' => 'required',
-        'categoria_id' => 'required',
-        'nombre' => 'required|unique:producto',
-        'precio_unit' => 'required',
-        'cantidad_por_caja' => 'required',
-        'foto' => 'required',
-        'punto_reorden' => 'required',
-        'cantidad_caja' => 'required',
-    ]);
+    public function guardarProducto(Request $request){
+        // Validar la solicitud
+        $request->validate([
+            'marca_id' => 'required',
+            'categoria_id' => 'required',
+            'nombre' => 'required|unique:producto',
+            'precio_unit' => 'required',
+            'cantidad_por_caja' => 'required',
+            'foto' => 'required',
+            'punto_reorden' => 'required',
+            'cantidad_cajas' => 'required',
+        ]);
 
-    try {
-        // Validación para comprobar si ya existe un producto con el mismo nombre
-        $nombre = $request->input('nombre');
-        $productoExistente = Producto::where('nombre', $nombre)->first();
+        try {
+            // Validación para comprobar si ya existe un producto con el mismo nombre
+            $nombre = $request->input('nombre');
+            $productoExistente = Producto::where('nombre', $nombre)->first();
 
-        if ($productoExistente) {
-            return response()->json(['error' => 'Ya existe un producto con el mismo nombre.'], 400);
+            if ($productoExistente) {
+                return response()->json(['error' => 'Ya existe un producto con el mismo nombre.'], 400);
+            }
+
+            // Crear un nuevo producto
+            $nuevoProducto = new Producto;
+            $nuevoProducto->marca_id = $request->input('marca_id');
+            $nuevoProducto->categoria_id = $request->input('categoria_id');
+            $nuevoProducto->nombre = $nombre;
+            $nuevoProducto->precio_unit = $request->input('precio_unit');
+            $nuevoProducto->cantidad_por_caja = $request->input('cantidad_por_caja');
+            $nuevoProducto->foto = $request->input('foto');
+            $nuevoProducto->punto_reorden = $request->input('punto_reorden');
+            $nuevoProducto->cantidad_cajas = $request->input('cantidad_cajas');
+            $nuevoProducto->save();
+
+            // Devolver una respuesta exitosa
+            return response()->json(['message' => 'Producto creado con éxito.'], 201);
+        } catch (Exception $e) {
+            // Manejar cualquier error y devolver una respuesta de error
+            return response()->json(['error' => 'Error al procesar la solicitud.'], 500);
         }
-
-        // Crear un nuevo producto
-        $nuevoProducto = new Producto;
-        $nuevoProducto->marca_id = $request->input('marca_id');
-        $nuevoProducto->categoria_id = $request->input('categoria_id');
-        $nuevoProducto->nombre = $nombre;
-        $nuevoProducto->precio_unit = $request->input('precio_unit');
-        $nuevoProducto->cantidad_por_caja = $request->input('cantidad_por_caja');
-        $nuevoProducto->foto = $request->input('foto');
-        $nuevoProducto->punto_reorden = $request->input('punto_reorden');
-        $nuevoProducto->cantidad_caja = $request->input('cantidad_caja');
-        $nuevoProducto->save();
-
-        // Devolver una respuesta exitosa
-        return response()->json(['message' => 'Producto creado con éxito.'], 201);
-    } catch (\Exception $e) {
-        // Manejar cualquier error y devolver una respuesta de error
-        return response()->json(['error' => 'Error al procesar la solicitud.'], 500);
     }
-}
 
+     //obtiene una producto en especifico
+     public function getProducto($id_producto) {
+        try {
+            $producto = Producto::findOrFail($id_producto);
+    
+            return response()->json(["data" => $producto, "status" => 200]);
+        } catch (Exception $e) {
+            print($e);
+            return response()->json(["mensaje" => "Error al obtener la marca", "status" => 500]);
+        }
+    }
 
     //actualiza la informacion de un producto en especifico
     public function updateProducto(Request $request, $id_producto) {
@@ -82,12 +92,17 @@ class ProductoController extends Controller
     //elimina un producto en especifico
     public function deleteProducto($id_producto){
         try {
-            $producto = Producto::findOrFail($id_producto);
+            $producto = Producto::find($id_producto);
+
+            if (!$producto) {
+                return response()->json(['error' => 'Producto no encontrado'], 404);
+            }
             $producto->delete();
 
-            return response()->json(['success' => true, 'message' => 'Producto eliminado con éxito'], 200);
-        } catch (\Exception $e) {
-            return response()->json(['success' => false, 'message' => 'Producto no encontrado'], 404);
+            return response()->json(['message' => 'Producto eliminado correctamente'], 404);
+        } catch (Exception $e) {
+            print($e);
+            return response()->json(['error' => 'Error al procesar la solicitud.'], 500);
         }
     }    
 }
